feat: save number of pages and new text extractor (#39)
Knowing how many pages are in a PDF can tell a lot about the contents. The text extracted of course says everything, but sometimes there are scans and short pdfs that are text heavy.

Also switches to pdfminer.six and uses a better way to get text out.

This doesn't expose the new data everywhere. Need to get this merged to first to get the data.

Closes #36